Enum erased_cells::CellValue

pub enum CellValue {
    UInt8(u8),
    UInt16(u16),
    UInt32(u32),
    UInt64(u64),
    Int8(i8),
    Int16(i16),
    Int32(i32),
    Int64(i64),
    Float32(f32),
    Float64(f64),
}

Value variants for each CellType

Variants

UInt8(u8)

UInt16(u16)

UInt32(u32)

UInt64(u64)

Int8(i8)

Int16(i16)

Int32(i32)

Int64(i64)

Float32(f32)

Float64(f64)

Implementations

impl CellValue

pub fn new<T: CellEncoding + Sized>(value: T) -> Self

Construct new CellValue from a statically known CellEncoding.

pub fn cell_type(&self) -> CellType

Get the CellType encoding self.

pub fn get<T: CellEncoding>(&self) -> Result<T>

Get the CellValue contents as a T.

Returns Ok(T) if the CellType of T is the same or wider than the encoded value, or Err(Error) if T is narrower.

pub fn convert(&self, cell_type: CellType) -> Result<Self>

Convert self into a variant with CellType cell_type equal to or wider than its current CellType.

Returns Ok(CellValue) if the CellType of cell_type is the same or wider than the encoded value, or Err(Error) if T is narrower.

pub fn unify(&self, other: &Self) -> (Self, Self)

Determines the smallest cell type that can contain self and other, and then converts values to that cell type and returns a tuple of the converted values, i.e. (convert(self), convert(other)).