Trait erased_cells::BufferOps

pub trait BufferOps {
    // Required methods
    fn from_vec<T: CellEncoding>(data: Vec<T>) -> Self;
    fn with_defaults(len: usize, ct: CellType) -> Self;
    fn fill(len: usize, value: CellValue) -> Self;
    fn fill_via<T, F>(len: usize, f: F) -> Self
       where T: CellEncoding,
             F: Fn(usize) -> T;
    fn len(&self) -> usize;
    fn cell_type(&self) -> CellType;
    fn get(&self, index: usize) -> CellValue;
    fn put(&mut self, idx: usize, value: CellValue) -> Result<()>;
    fn convert(&self, cell_type: CellType) -> Result<Self>
       where Self: Sized;
    fn min_max(&self) -> (CellValue, CellValue);
    fn to_vec<T: CellEncoding>(self) -> Result<Vec<T>>;

    // Provided method
    fn is_empty(&self) -> bool { ... }
}

Required Methods

fn from_vec<T: CellEncoding>(data: Vec<T>) -> Self

Construct a CellBuffer from a Vec<T>.

fn with_defaults(len: usize, ct: CellType) -> Self

Construct a CellBuffer of given len length and ct CellType

All cells will be filled with the CellType’s corresponding default value.

fn fill(len: usize, value: CellValue) -> Self

Create a buffer of size len with all values value.

fn fill_via<T, F>(len: usize, f: F) -> Selfwhere T: CellEncoding, F: Fn(usize) -> T,

Fill a buffer of size len with values from a closure.

First parameter of the closure is the current index.

fn len(&self) -> usize

Get the length of the buffer.

fn cell_type(&self) -> CellType

Get the cell type of the encoded value.

fn get(&self, index: usize) -> CellValue

Get the CellValue at index idx.

Panics of idx is outside of [0, len()).

fn put(&mut self, idx: usize, value: CellValue) -> Result<()>

Store value at position idx.

fn convert(&self, cell_type: CellType) -> Result<Self>where Self: Sized,

Create a new CellBuffer whereby all CellValues are converted to cell_type.

Returns Ok(CellBuffer) if conversion is possible, and Err(Error) if contained values cannot fit in cell_type without clamping.

fn min_max(&self) -> (CellValue, CellValue)

Compute the minimum and maximum values the buffer.

fn to_vec<T: CellEncoding>(self) -> Result<Vec<T>>

Convert self into a Vec<T>.

Provided Methods

fn is_empty(&self) -> bool

Determine if the buffer has zero values in it.